Move reference data module to data.
